requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
audioread==3.0.1
azure-core==1.35.1
azure-storage-blob==12.26.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
boto3==1.40.36
botocore==1.40.36
Brotli==1.1.0
cachetools==5.5.2
canopen==2.4.1
celery==5.4.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.7
cryptography==45.0.7
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.8
django-appconf==1.1.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-cotton==2.1.3
django-countries==7.6.1
django-filter==25.1
django-imagekit==5.0.0
django-import-export==4.3.9
django-jazzmin==3.0.1
django-json-widget==2.0.3
django-modeltranslation==0.19.16
django-phonenumber-field==8.1.0
django-pwa==2.0.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
edge-tts==7.2.3
einops==0.8.1
et_xmlfile==2.0.0
Faker==37.6.0
faker-file==0.18.4
filelock==3.19.1
fonttools==4.60.0
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gdown==5.2.0
gensim==4.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==1.44.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
imgkit==1.2.3
inflection==0.5.1
iniconfig==2.1.0
invoke==2.2.0
iso3166==2.1.1
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mock==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
narwhals==2.5.0
networkx==3.5
nltk==3.9.1
numba==0.62.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
odfpy==1.4.1
opcua==0.98.13
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
paramiko==4.0.0
pathlib_abc==0.1.1
pathy==0.11.0
pdf2image==1.17.0
pdfkit==1.0.0
pgvector==0.4.1
phonenumbers==9.0.13
pilkit==3.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
plotly-geo==1.0.0
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
psycopg==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pymodbus==3.11.2
PyNaCl==1.6.0
PyPDF2==3.0.1
pyphen==0.17.2
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.2
pytest-cov==7.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
python-can==4.6.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
python-stdnum==2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
qrbill==1.2.0
qrcode==8.2
redis==5.0.8
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rlPyCairo==0.4.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.13.1
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
sqlparse==0.5.3
standard-imghdr==3.13.0
svglib==1.6.0
svgwrite==1.4.3
sympy==1.14.0
tablib==3.8.0
tabulate==0.9.0
textaugment==2.0.0
textblob==0.17.1
threadpoolctl==3.6.0
tika==3.1.0
timm==1.0.19
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.1
triton==3.4.0
typer==0.19.1
typer-cli==0.19.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.11.0
wrapt==1.17.3
xlsxwriter==3.2.9
xml2epub==2.6.8
yarl==1.20.1
zopfli==0.2.3.post1
